refactor: use jasmine-ts for unit testing
