+++ jasmine.config.json
@@ -1,0 +1,12 @@
+{
+  "reporters": [
+    {
+      "name": "jasmine-spec-reporter#SpecReporter",
+      "options": {
+        "displayStacktrace": "all"
+      }
+    }
+  ],
+  "spec_dir": "test",
+  "spec_files": ["**/*[sS]pec.ts"]
+}
