--- src/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../tsconfig-base.json",
-  "compilerOptions": {
-    "composite": true,
-    "rootDir": "./",
-    "outDir": "../dist"
-  }
-}
