--- test/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../tsconfig-base.json",
-  "references": [
-    { "path": "../src" }
-  ],
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./compiled"
-  }
-}
